moved contract from boa to boa.vyper
